+++ CHANGES.rst
@@ -6,7 +6,8 @@ There's a frood who really knows where his towel is.
 2.1b3 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+- Fix Plone 5.1 error when index Tile RichText content.
+  [ale-rt, rodfersou]
 
 
 2.1b2 (2018-10-04)